+++ src/components/Popover/Popover.types.ts
@@ -22,6 +22,7 @@ export type Props = Pick<
 	| "disableHideAnimation"
 	| "disableContentHover"
 	| "disableCloseOnOutsideClick"
+	| "autoFocus"
 	| "containerRef"
 	| "initialFocusRef"
 	| "originCoordinates"